make multiselectfield init more robust
In a newer version of chafield, the max length is no longer required and the validator may be missing. This code should fix that. Minor note I want to add: the init code of Field actually caused the maxlen validator to be last, not first, so you'd need to always take [-1] instead of [0] or you'd overwrite user-supplied validators (but that's also fixed here)